LoginController logic refactorization
